feat: Añadir gestión de eliminación de embarques por usuario
- Se actualiza la lógica en EmbarquesOfflineService para incluir un nuevo campo 'deletedByUser' que indica si el embarque fue eliminado por el usuario.
- Se modifica la verificación de eliminación en ListaEmbarques y NuevoEmbarque para considerar tanto 'deleted' como 'deletedByUser', asegurando que solo se eliminen los registros que han sido marcados por el usuario.

+++ src/views/Embarques/ListaEmbarques.vue
@@ -385,7 +385,7 @@ export default {
             const embarqueRef = doc(db, 'embarques', record.id);
             const snapshot = await getDoc(embarqueRef);
 
-            if (record.deleted) {
+            if (record.deleted && record.deletedByUser) {
               if (snapshot.exists()) {
                 await deleteDoc(embarqueRef);
               }
